+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-- 2023-04-05: Chiron experimental plugins.
+- 2023-04-05:
+  - added Chiron experimental plugins.
+  - fixed sort by reversed value in terms query builder.
 
 ## [0.0.10] - 2023-03-29